+++ artiq/language/environment.py
@@ -321,6 +321,18 @@ class HasEnvironment:
         as ``slice(*sub_tuple)`` (multi-dimensional slicing)."""
         self.__dataset_mgr.mutate(key, index, value)
 
+    @rpc(flags={"async"})
+    def append_to_dataset(self, key, value):
+        """Append a value to a dataset.
+
+        The target dataset must be a list (i.e. support ``append()``), and must
+        have previously been set from this experiment.
+
+        The broadcast/persist/archive mode of the given key remains unchanged
+        from when the dataset was last set. Appended values are transmitted
+        efficiently as incremental modifications in broadcast mode."""
+        self.__dataset_mgr.append_to(key, value)
+
     def get_dataset(self, key, default=NoDefault, archive=True):
         """Returns the contents of a dataset.